# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/v2/generation/text-to-image-with-ip": {
            "post": {
                "summary": "Text To Img With Ip",
                "operationId": "text_to_img_with_ip_v2_generation_text_to_image_with_ip_post",
                "parameters": [
                    {
                        "name": "accept",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Parameter to overvide 'Accept' header, 'image/png' for output bytes",
                            "title": "Accept"
                        },
                        "description": "Parameter to overvide 'Accept' header, 'image/png' for output bytes"
                    },
                    {
                        "name": "accept",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Accept"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Text2ImgRequestWithPrompt"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PNG bytes if request's 'Accept' header is 'image/png', otherwise JSON",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GeneratedImageResult"
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/AsyncJobResponse"
                                        }
                                    ],
                                    "title": "Response Text To Img With Ip V2 Generation Text To Image With Ip Post"
                                },
                                "example": [
                                    {
                                        "base64": "...very long string...",
                                        "seed": "1050625087",
                                        "finish_reason": "SUCCESS"
                                    }
                                ]
                            },
                            "application/json async": {
                                "example": {
                                    "job_id": 1,
                                    "job_type": "Text to Image"
                                }
                            },
                            "image/png": {
                                "example": "PNG bytes, what did you expect?"
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AsyncJobResponse": {
                "properties": {
                    "job_id": {
                        "type": "string",
                        "title": "Job Id",
                        "description": "Job ID"
                    },
                    "job_type": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/TaskType"
                            }
                        ],
                        "description": "Job type"
                    },
                    "job_stage": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/AsyncJobStage"
                            }
                        ],
                        "description": "Job running stage"
                    },
                    "job_progress": {
                        "type": "integer",
                        "title": "Job Progress",
                        "description": "Job running progress, 100 is for finished."
                    },
                    "job_status": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Job Status",
                        "description": "Job running status in text"
                    },
                    "job_step_preview": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Job Step Preview",
                        "description": "Preview image of generation steps at current time, as base64 image"
                    },
                    "job_result": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/GeneratedImageResult"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Job Result",
                        "description": "Job generation result"
                    }
                },
                "type": "object",
                "required": [
                    "job_id",
                    "job_type",
                    "job_stage",
                    "job_progress"
                ],
                "title": "AsyncJobResponse"
            },
            "AsyncJobStage": {
                "type": "string",
                "enum": [
                    "WAITING",
                    "RUNNING",
                    "SUCCESS",
                    "ERROR"
                ],
                "title": "AsyncJobStage"
            },
...
            "ControlNetType": {
                "type": "string",
                "enum": [
                    "ImagePrompt",
                    "FaceSwap",
                    "PyraCanny",
                    "CPDS"
                ],
                "title": "ControlNetType"
            },
...
            "GeneratedImageResult": {
                "properties": {
                    "base64": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Base64",
                        "description": "Image encoded in base64, or null if finishReasen is not 'SUCCESS', only return when request require base64"
                    },
                    "url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Url",
                        "description": "Image file static serve url, or null if finishReasen is not 'SUCCESS'"
                    },
                    "seed": {
                        "type": "string",
                        "title": "Seed",
                        "description": "The seed associated with this image"
                    },
                    "finish_reason": {
                        "$ref": "#/components/schemas/GenerationFinishReason"
                    }
                },
                "type": "object",
                "required": [
                    "base64",
                    "url",
                    "seed",
                    "finish_reason"
                ],
                "title": "GeneratedImageResult"
            },
            "GenerationFinishReason": {
                "type": "string",
                "enum": [
                    "SUCCESS",
                    "QUEUE_IS_FULL",
                    "USER_CANCEL",
                    "ERROR"
                ],
                "title": "GenerationFinishReason"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ImagePromptJson": {
                "properties": {
                    "cn_img": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cn Img",
                        "description": "Input image for image prompt as base64"
                    },
                    "cn_stop": {
                        "anyOf": [
                            {
                                "type": "number",
                                "maximum": 1.0,
                                "minimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cn Stop",
                        "description": "Stop at for image prompt, 0 for default value",
                        "default": 0
                    },
                    "cn_weight": {
                        "anyOf": [
                            {
                                "type": "number",
                                "maximum": 2.0,
                                "minimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cn Weight",
                        "description": "Weight for image prompt, 0 for default value",
                        "default": 0
                    },
                    "cn_type": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/ControlNetType"
                            }
                        ],
                        "description": "ControlNet type for image prompt",
                        "default": "ImagePrompt"
                    }
                },
                "type": "object",
                "title": "ImagePromptJson"
            },
...
            "Lora": {
                "properties": {
                    "model_name": {
                        "type": "string",
                        "title": "Model Name"
                    },
                    "weight": {
                        "type": "number",
                        "maximum": 2.0,
                        "minimum": -2.0,
                        "title": "Weight",
                        "default": 0.5
                    }
                },
                "type": "object",
                "required": [
                    "model_name"
                ],
                "title": "Lora"
            },
...
            "PerfomanceSelection": {
                "type": "string",
                "enum": [
                    "Speed",
                    "Quality",
                    "Extreme Speed"
                ],
                "title": "PerfomanceSelection"
            },
...
            "TaskType": {
                "type": "string",
                "enum": [
                    "Text to Image",
                    "Image Upscale or Variation",
                    "Image Inpaint or Outpaint",
                    "Image Prompt",
                    "Not Found"
                ],
                "title": "TaskType"
            },
...
            "Text2ImgRequestWithPrompt": {
                "properties": {
                    "prompt": {
                        "type": "string",
                        "title": "Prompt",
                        "default": ""
                    },
                    "negative_prompt": {
                        "type": "string",
                        "title": "Negative Prompt",
                        "default": ""
                    },
                    "style_selections": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Style Selections",
                        "default": [
                            "Fooocus V2",
                            "Fooocus Enhance",
                            "Fooocus Sharp"
                        ]
                    },
                    "performance_selection": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/PerfomanceSelection"
                            }
                        ],
                        "default": "Speed"
                    },
                    "aspect_ratios_selection": {
                        "type": "string",
                        "title": "Aspect Ratios Selection",
                        "default": "800*1200"
                    },
                    "image_number": {
                        "type": "integer",
                        "maximum": 32.0,
                        "minimum": 1.0,
                        "title": "Image Number",
                        "description": "Image number",
                        "default": 1
                    },
                    "image_seed": {
                        "type": "integer",
                        "title": "Image Seed",
                        "description": "Seed to generate image, -1 for random",
                        "default": -1
                    },
                    "sharpness": {
                        "type": "number",
                        "maximum": 30.0,
                        "minimum": 0.0,
                        "title": "Sharpness",
                        "default": 2.0
                    },
                    "guidance_scale": {
                        "type": "number",
                        "maximum": 30.0,
                        "minimum": 1.0,
                        "title": "Guidance Scale",
                        "default": 4.0
                    },
                    "base_model_name": {
                        "type": "string",
                        "title": "Base Model Name",
                        "default": "juggernautXL_v8Rundiffusion.safetensors"
                    },
                    "refiner_model_name": {
                        "type": "string",
                        "title": "Refiner Model Name",
                        "default": "None"
                    },
                    "refiner_switch": {
                        "type": "number",
                        "maximum": 1.0,
                        "minimum": 0.1,
                        "title": "Refiner Switch",
                        "description": "Refiner Switch At",
                        "default": 0.5
                    },
                    "loras": {
                        "items": {
                            "$ref": "#/components/schemas/Lora"
                        },
                        "type": "array",
                        "title": "Loras",
                        "default": [
                            {
                                "model_name": "sd_xl_offset_example-lora_1.0.safetensors",
                                "weight": 0.1
                            }
                        ]
                    },
                    "advanced_params": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/AdvancedParams"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "default": {
                            "adaptive_cfg": 7.0,
                            "adm_scaler_end": 0.3,
                            "adm_scaler_negative": 0.8,
                            "adm_scaler_positive": 1.5,
                            "canny_high_threshold": 128,
                            "canny_low_threshold": 64,
                            "controlnet_softness": 0.25,
                            "debugging_cn_preprocessor": false,
                            "debugging_inpaint_preprocessor": false,
                            "disable_preview": false,
                            "freeu_b1": 1.01,
                            "freeu_b2": 1.02,
                            "freeu_enabled": false,
                            "freeu_s1": 0.99,
                            "freeu_s2": 0.95,
                            "inpaint_disable_initial_latent": false,
                            "inpaint_engine": "v1",
                            "inpaint_erode_or_dilate": 0,
                            "inpaint_respective_field": 1.0,
                            "inpaint_strength": 1.0,
                            "invert_mask_checkbox": false,
                            "mixing_image_prompt_and_inpaint": false,
                            "mixing_image_prompt_and_vary_upscale": false,
                            "overwrite_height": -1,
                            "overwrite_step": -1,
                            "overwrite_switch": -1,
                            "overwrite_upscale_strength": -1.0,
                            "overwrite_vary_strength": -1.0,
                            "overwrite_width": -1,
                            "refiner_swap_method": "joint",
                            "sampler_name": "dpmpp_2m_sde_gpu",
                            "scheduler_name": "karras",
                            "skipping_cn_preprocessor": false
                        }
                    },
                    "require_base64": {
                        "type": "boolean",
                        "title": "Require Base64",
                        "description": "Return base64 data of generated image",
                        "default": false
                    },
                    "async_process": {
                        "type": "boolean",
                        "title": "Async Process",
                        "description": "Set to true will run async and return job info for retrieve generataion result later",
                        "default": false
                    },
                    "webhook_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Webhook Url",
                        "description": "Optional URL for a webhook callback. If provided, the system will send a POST request to this URL upon task completion or failure. This allows for asynchronous notification of task status.",
                        "default": ""
                    },
                    "image_prompts": {
                        "items": {
                            "$ref": "#/components/schemas/ImagePromptJson"
                        },
                        "type": "array",
                        "title": "Image Prompts",
                        "default": []
                    }
                },
                "type": "object",
                "title": "Text2ImgRequestWithPrompt"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}